refactor(sdk): split shielded mod.rs into decrypt, types, and sync files
Move logical groups out of the monolithic mod.rs:
- decrypt.rs: try_decrypt_note + MIN_ENCRYPTED_NOTE_LEN constant
- types.rs: ShieldedSyncConfig, DecryptedNote, ShieldedSyncResult
- sync.rs: sync_shielded_notes + fetch_chunk

mod.rs now just re-exports public items.
