+++ packages/rs-sdk/src/platform/shielded/nullifier_sync/types.rs
