+++ packages/rs-sdk/src/platform/shielded/sync.rs
@@ -1,134 +1,19 @@
-//! Shielded note sync and trial decryption utilities.
-//!
-//! This module provides:
-//! - [`try_decrypt_note`]: compact trial decryption on a single encrypted note
-//! - [`sync_shielded_notes`]: end-to-end sync that fetches encrypted notes from
-//!   the network in parallel and performs trial decryption
+//! Shielded note sync: parallel chunk fetching with trial decryption.
 
+use super::decrypt::try_decrypt_note;
+use super::types::{DecryptedNote, ShieldedSyncConfig, ShieldedSyncResult};
 use crate::error::Error;
 use crate::platform::Fetch;
 use crate::Sdk;
 use drive_proof_verifier::types::ShieldedEncryptedNotesQuery;
 use drive_proof_verifier::types::{ShieldedEncryptedNote, ShieldedEncryptedNotes};
 use futures::stream::{FuturesUnordered, StreamExt};
-use grovedb_commitment_tree::{
-    try_compact_note_decryption, CompactAction, DashMemo, EphemeralKeyBytes,
-    ExtractedNoteCommitment, Note, Nullifier, OrchardDomain, PaymentAddress,
-    PreparedIncomingViewingKey, COMPACT_NOTE_SIZE,
-};
+use grovedb_commitment_tree::PreparedIncomingViewingKey;
 use rs_dapi_client::RequestSettings;
 use std::collections::BTreeMap;
 use std::future::Future;
 use std::pin::Pin;
 use tracing::debug;
-
-/// Minimum length of the `encrypted_note` field for compact trial decryption.
-///
-/// The `encrypted_note` field layout is:
-///   `epk(32) || enc_ciphertext(104) || out_ciphertext(80)` = 216 bytes
-///
-/// For compact decryption we need at least `epk(32) + COMPACT_NOTE_SIZE` bytes
-/// of the enc_ciphertext.
-const MIN_ENCRYPTED_NOTE_LEN: usize = 32 + COMPACT_NOTE_SIZE;
-
-/// Default maximum number of chunk queries in flight at once.
-const DEFAULT_MAX_CONCURRENT: usize = 4;
-
-// ---------------------------------------------------------------------------
-// Trial decryption
-// ---------------------------------------------------------------------------
-
-/// Attempt compact trial decryption on a [`ShieldedEncryptedNote`].
-///
-/// The [`ShieldedEncryptedNote`] struct (from proof verification) has three
-/// separate fields:
-///   - `cmx`: note commitment (32 bytes)
-///   - `nullifier`: nullifier (32 bytes) — used for Rho derivation
-///   - `encrypted_note`: `epk(32) || enc_ciphertext(104) || out_ciphertext(80)`
-///
-/// Returns `Some((note, address))` if the note decrypts successfully under the
-/// given incoming viewing key, or `None` if it does not belong to the viewer
-/// (including dummy/padding notes).
-pub fn try_decrypt_note(
-    ivk: &PreparedIncomingViewingKey,
-    encrypted_note: &ShieldedEncryptedNote,
-) -> Option<(Note, PaymentAddress)> {
-    let data = &encrypted_note.encrypted_note;
-    if data.len() < MIN_ENCRYPTED_NOTE_LEN {
-        return None;
-    }
-
-    // Parse nullifier from the dedicated field (32 bytes)
-    let nf_bytes: [u8; 32] = encrypted_note.nullifier.as_slice().try_into().ok()?;
-    let nf = Nullifier::from_bytes(&nf_bytes).into_option()?;
-
-    // Parse cmx from the dedicated field (32 bytes)
-    let cmx_bytes: [u8; 32] = encrypted_note.cmx.as_slice().try_into().ok()?;
-    let cmx = ExtractedNoteCommitment::from_bytes(&cmx_bytes).into_option()?;
-
-    // Parse ephemeral public key (first 32 bytes of encrypted_note)
-    let epk_bytes: [u8; 32] = data[0..32].try_into().ok()?;
-
-    // Parse compact ciphertext (first COMPACT_NOTE_SIZE bytes of enc_ciphertext,
-    // starting at byte 32)
-    let enc_compact: [u8; COMPACT_NOTE_SIZE] = data[32..32 + COMPACT_NOTE_SIZE].try_into().ok()?;
-
-    // Build CompactAction and OrchardDomain for trial decryption
-    let compact = CompactAction::from_parts(nf, cmx, EphemeralKeyBytes(epk_bytes), enc_compact);
-    let domain = OrchardDomain::<DashMemo>::for_compact_action(&compact);
-
-    try_compact_note_decryption(&domain, ivk, &compact)
-}
-
-// ---------------------------------------------------------------------------
-// Shielded note sync
-// ---------------------------------------------------------------------------
-
-/// Configuration for [`sync_shielded_notes`].
-pub struct ShieldedSyncConfig {
-    /// Maximum number of chunk queries in flight at once (default: 4).
-    pub max_concurrent: usize,
-    /// Request settings forwarded to each individual fetch call.
-    pub request_settings: RequestSettings,
-}
-
-impl Default for ShieldedSyncConfig {
-    fn default() -> Self {
-        Self {
-            max_concurrent: DEFAULT_MAX_CONCURRENT,
-            request_settings: RequestSettings::default(),
-        }
-    }
-}
-
-/// A note that was successfully decrypted (belongs to the viewer).
-pub struct DecryptedNote {
-    /// Global position of this note in the commitment tree.
-    pub position: u64,
-    /// The decrypted Orchard note (contains value, rseed, rho).
-    pub note: Note,
-    /// The recipient payment address.
-    pub address: PaymentAddress,
-    /// The nullifier (32 bytes).
-    pub nullifier: [u8; 32],
-    /// The note commitment (32 bytes).
-    pub cmx: [u8; 32],
-}
-
-/// Result of [`sync_shielded_notes`].
-pub struct ShieldedSyncResult {
-    /// Notes that successfully decrypted (belong to the viewer).
-    pub decrypted_notes: Vec<DecryptedNote>,
-    /// All raw encrypted notes fetched, in tree order.
-    /// Useful for updating a local commitment tree.
-    pub all_notes: Vec<ShieldedEncryptedNote>,
-    /// Next chunk-aligned index to resume syncing from.
-    pub next_start_index: u64,
-    /// Total number of notes scanned in this sync.
-    pub total_notes_scanned: u64,
-    /// Platform block height at the time of the most recent chunk response.
-    pub block_height: u64,
-}
 
 /// Fetch all shielded encrypted notes starting from `start_index`, query
 /// multiple nodes in parallel, and perform trial decryption.
